bug fix - [REST API] Self-check > /api/v1/create_selfcheck

+++ src/main/java/oss/fosslight/service/impl/ApiSelfCheckServiceImpl.java
@@ -33,7 +33,7 @@ public class ApiSelfCheckServiceImpl implements ApiSelfCheckService {
 		
 		apiSelfcheckMapper.createSelfCheck(paramMap);
 		
-		BigInteger prjId = (BigInteger) paramMap.get("prjId");
+		Long prjId = (Long) paramMap.get("prjId");
 		result.put("prjId", prjId.toString());
 		
 		return result;
